Sum of Two numbers represented by list - First Commit
You have two numbers represented by a linked list where each node contains a single digit. The digits are store in reverse order. such that 1's digit is at the head of the list. Write a function that adds the two numbers and returns the sum as a linked list.